requirements.txt
# INFOCAM - install from project root: pip install -r requirements.txt
#
# For better inference (GPU): after the above, install PyTorch with CUDA:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# Then restart the app; models will use GPU and video will run smoother.

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0
asyncpg==0.29.0
ultralytics==8.0.196
opencv-python==4.8.1.78
numpy==1.24.3
pydantic==2.5.0
email-validator==2.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
